+++ src/controllers/comment.js
@@ -1,7 +1,24 @@
-import { prisma } from "./../../lib/prisma.js";
+import { prisma } from "./../lib/prisma.js";
 
 export const createNewComment = async (req, res) => {
     try {
+        const authorId = req.body.authorId;
+        const blogId = req.body.blogId
+        if (!authorId || !blogId) {
+            return res.status(400).json({ message: 'Author ID and Blog ID is required' });
+        }
+
+        const user = await prisma.users.findUnique({
+            where: {
+                id: authorId
+            }
+        });
+        if (!user) {
+            return res.status(404).json({
+                message: 'Author not found'
+            });
+        }
+
         const comment = await prisma.comment.create({
             data: {
                 content: req.body.content,
@@ -16,13 +33,20 @@ export const createNewComment = async (req, res) => {
         })
     } catch (error) {
         console.error('Error creating new comment:', error);
-        res.status(500).json({ message: 'Internal server error' });
+        res.status(400).json({ message: 'Bad Request' });
     }
 }
 
 export const updateComment = async (req, res) => {
     const commentId = parseInt(req.params.commentId);
     try {
+        const findComment = await prisma.comment.findUnique({
+            where: { id: commentId }
+        });
+        if (!findComment) {
+            return res.status(404).json({ message: 'Comment not found' });
+        }
+
         const comment = await prisma.comment.update({
             where: {
                 id: commentId
@@ -36,13 +60,21 @@ export const updateComment = async (req, res) => {
         })
     } catch (error) {
         console.error('Error updating comment:', error);
-        res.status(404).json({ message: 'Tag not found' });
+        res.status(500).json({ message: 'Internal Server Error' });
     }
 }
 
 export const deleteComment = async (req, res) => {
     const commentId = parseInt(req.params.commentId);
     try {
+        const findComment = await prisma.comment.findUnique({
+            where: { id: commentId }
+        })
+
+        if (!findComment) {
+            return res.status(404).json({ message: 'Comment not found' });
+        }
+
         const comment = await prisma.comment.delete({
             where: {
                 id: commentId
@@ -55,6 +87,6 @@ export const deleteComment = async (req, res) => {
         })
     } catch (error) {
         console.error('Error deleting comment:', error);
-        res.status(404).json({ message: 'Comment not found' });
+        res.status(500).json({ message: 'Internal Server Error' });
     }
 }
